+++ R/whit_loggit.R
@@ -29,7 +29,7 @@
 #' }
 #'
 #' @export
-with_loggit <- function(exp, logfile = get_logfile(), echo = TRUE, log_level = get_log_level()) {
+with_loggit <- function(exp, logfile = get_logfile(), echo = get_echo(), log_level = get_log_level()) {
   log_level <- convert_lvl_input(log_level)
   withCallingHandlers(
     exp,